Pantalla registro adoptantes acabada
